Trading-Bot/requirements.txt
python-binance
loguru
typer
rich
python-dotenv